% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{sportsdataverse_upload}
\alias{sportsdataverse_upload}
\title{Upload to sportsdataverse release}
\usage{
sportsdataverse_upload(...)
}
\arguments{
\item{...}{currently not used}

\item{files}{vector of filepaths to upload}

\item{tag}{release name}

\item{pkg_function}{related package function name}

\item{repo}{repository to upload to, default: \code{"sportsdataverse/sportsdataverse-data"}}

\item{overwrite}{If \code{TRUE} (the default) existing files will be overwritten}
}
\description{
Upload to sportsdataverse release
}
