% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{sportsdataverse_upload}
\alias{sportsdataverse_upload}
\title{Upload to sportsdataverse release}
\usage{
sportsdataverse_upload(...)
}
\arguments{
\item{...}{other args passed to \code{piggyback::pb_upload()}}

\item{files}{vector of filepaths to upload}

\item{tag}{release name}
}
\description{
Upload to sportsdataverse release
}
